+++ lib/main.dart
@@ -2,9 +2,12 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:my_app/views/register_view.dart';
+import 'package:my_app/views/reports_view.dart';
+import 'package:my_app/views/verify_email.dart';
 import 'package:my_app/views/verify_email.dart';
 import 'firebase_options.dart';
 import '../views/login_view.dart';
+import '../views/reports_view.dart' as ReportsView;
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -18,8 +21,10 @@ void main() {
       routes: {
         '/login/': (context) => const LoginView(),
         '/register/': (context) => const RegisterView(),
-        '/reports/': (context) => const ReportsView(),
+        '/reports/': (context) => const ReportsView.ReportsView(),
+        '/verifyEmail/': (context) => const VerifyEmailView()
       },
+      debugShowCheckedModeBanner: false,
     ),
   );
 }
@@ -34,91 +39,23 @@ class HomePage extends StatelessWidget {
         options: DefaultFirebaseOptions.currentPlatform,
       ),
       builder: (context, snapshot) {
-        switch (snapshot.connectionState) {
-          case ConnectionState.done:
-            final user = FirebaseAuth.instance.currentUser;
-            if (user != null) {
-              if (user.emailVerified) {
-                return const ReportsView();
-              } else {
-                return const VerifyEmailView();
-              }
-            } else {
-              return const LoginView();
-            }
+        // switch (snapshot.connectionState) {
+        //   case ConnectionState.done:
+        //     final user = FirebaseAuth.instance.currentUser;
+        //     if (user != null) {
+        //       if (user.emailVerified) {
+        //         return const ReportsView();
+        //       } else {
+        //         return const VerifyEmailView();
+        //       }
+        //     } else {
+        return const LoginView();
+        //     }
 
-          default:
-            return const CircularProgressIndicator();
-        }
+        //   default:
+        //     return const CircularProgressIndicator();
+        // }
       },
     );
   }
-}
-
-enum MenuAction { logout }
-
-class ReportsView extends StatefulWidget {
-  const ReportsView({super.key});
-
-  @override
-  State<ReportsView> createState() => _ReportsViewState();
-}
-
-class _ReportsViewState extends State<ReportsView> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Reports'),
-        actions: [
-          PopupMenuButton<MenuAction>(
-            onSelected: (value) async {
-              switch (value) {
-                case MenuAction.logout:
-                  final shouldLogout = await showLogOutDialog(context);
-                  if (shouldLogout) {
-                    await FirebaseAuth.instance.signOut();
-                    Navigator.of(context)
-                        .pushNamedAndRemoveUntil('/login/', (_) => false);
-                  }
-                  break;
-              }
-            },
-            itemBuilder: (context) {
-              return const [
-                PopupMenuItem<MenuAction>(
-                    value: MenuAction.logout, child: Text('Logout'))
-              ];
-            },
-          )
-        ],
-      ),
-      body: const Text('namskara reports'),
-    );
-  }
-}
-
-Future<bool> showLogOutDialog(BuildContext context) {
-  return showDialog<bool>(
-    context: context,
-    builder: (context) {
-      return AlertDialog(
-        title: const Text('Sign out'),
-        content: const Text('Are you sure you want to sign out?'),
-        actions: [
-          TextButton(
-              onPressed: () {
-                Navigator.of(context).pop(false);
-              },
-              child: const Text('Cancel')),
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop(true);
-            },
-            child: const Text('Log out'),
-          ),
-        ],
-      );
-    },
-  ).then((value) => value ?? false);
 }
